add sp tensor

+++ main_traffic4cast.py
@@ -70,7 +70,8 @@ Lk_sp = sp.coo_matrix(Lk)
 Lk_spt = tf.sparse_transpose(tf.SparseTensor(
     indices=np.array([Lk_sp.row, Lk_sp.col]).T,
     values=Lk_sp.data,
-    dense_shape=Lk_sp.shape))
+    dense_shape=Lk_sp.shape,
+    dtype=tf.float32))
 
 tf.add_to_collection(name='graph_kernel', value=Lk_spt)
